requirements_lock.txt
pandas==3.0.2
numpy==2.4.4
matplotlib==3.10.8
statsmodels==0.14.6
linearmodels==7.0
pyreadr==0.5.4
pyarrow==23.0.1
requests==2.33.1
sgfmill==1.1.1
scikit-learn==1.8.0
osfclient==0.0.5
Pillow==12.2.0
